requirements.txt
boto3==1.28.2
Flask==1.1.2
MarkupSafe==1.1.1
itsdangerous==2.0.1
werkzeug==2.0.3
Jinja2==2.11.2
click==7.1.2
mysql-connector-python==8.1.0
